web_ad_service/requirements.txt
# Web Ad Service Requirements

# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# AI/ML dependencies
google-generativeai==0.3.2
torch==2.1.0
numpy==1.24.3
scipy==1.11.4

# Web scraping and processing
urllib3==2.1.0
html5lib==1.1

# Database and storage
chromadb==0.4.18
elasticsearch==8.11.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
pathlib2==2.3.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: For enhanced text processing
nltk==3.8.1
spacy==3.7.2

